% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frsr-package.R
\docType{package}
\name{frsr-package}
\alias{frsr}
\alias{frsr-package}
\title{frsr: Imports Saved Data from the Family Resource Survey}
\description{
Functions to easily import microdata from Family Resources Survey datasets that have been downloaded from the UK Data Service. When provided with a folder where the FRS datasets are saved (including in sub-folders), ehsr will identify and import data for the years you specify, distinguishing between the household and housing stock datasets. It will also standardise variable names and value labels where these have evolved over time.
}
\author{
\strong{Maintainer}: James Gleeson \email{jimgleeson@gmail.com}

}
\keyword{internal}
