% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_adult}
\alias{import_adult}
\title{Import adult data, resulting in a dataset with mostly household variables
but one row for every adult. Only imports selected variables.}
\usage{
import_adult(folder, years)
}
\arguments{
\item{folder}{A folder containing FRS downloaded from UKDS and unzipped}

\item{years}{A list of years}
}
\description{
Import adult data, resulting in a dataset with mostly household variables
but one row for every adult. Only imports selected variables.
}
