% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_bu}
\alias{import_bu}
\title{Import benefit unit data}
\usage{
import_bu(folder, years)
}
\arguments{
\item{folder}{A folder containing FRS downloaded from UKDS and unzipped}

\item{years}{A list of years}
}
\description{
Import benefit unit data
}
