% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_child}
\alias{import_child}
\title{Import child data, resulting in a dataset with mostly household variables
but one row for every child Only imports selected variables.}
\usage{
import_child(folder, years)
}
\arguments{
\item{folder}{A folder containing FRS downloaded from UKDS and unzipped}

\item{years}{A list of years}
}
\description{
Import child data, resulting in a dataset with mostly household variables
but one row for every child Only imports selected variables.
}
