% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_hh}
\alias{import_hh}
\title{Import Family Resources Survey data}
\usage{
import_hh(folder, years)
}
\arguments{
\item{folder}{A folder containing FRS downloaded from UKDS and unzipped}

\item{years}{A list of years}
}
\description{
Import Family Resources Survey data
}
